+++ src/main/java/Pages/MainPage/MainPage.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.PageObject.MainPage;
+package Pages.MainPage;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
@@ -12,6 +12,27 @@ public class MainPage
     {
         this.chDriver = chDriver;
     }
+    private By[] locatorsHeaderQuestionAboutImportant =
+            {
+                    By.xpath(".//div[@aria-controls='accordion__panel-0']"),By.xpath(".//div[@aria-controls='accordion__panel-1']"),
+                    By.xpath(".//div[@aria-controls='accordion__panel-2']"), By.xpath(".//div[@aria-controls='accordion__panel-3']"),
+                    By.xpath(".//div[@aria-controls='accordion__panel-4']"),By.xpath(".//div[@aria-controls='accordion__panel-5']"),
+                    By.xpath(".//div[@aria-controls='accordion__panel-6']"), By.xpath(".//div[@aria-controls='accordion__panel-7']")
+            }; // массив локаторов кнопок с вопросами, от 0 до 7
+    private By[] locatorsBodyQuestionAboutImportant =
+            {
+                    By.xpath(".//div[@aria-labelledby='accordion__heading-0']/p"), By.xpath(".//div[@aria-labelledby='accordion__heading-1']/p"),
+                    By.xpath(".//div[@aria-labelledby='accordion__heading-2']/p"), By.xpath(".//div[@aria-labelledby='accordion__heading-3']/p"),
+                    By.xpath(".//div[@aria-labelledby='accordion__heading-4']/p"), By.xpath(".//div[@aria-labelledby='accordion__heading-5']/p"),
+                    By.xpath(".//div[@aria-labelledby='accordion__heading-6']/p"), By.xpath(".//div[@aria-labelledby='accordion__heading-7']/p")
+            }; //массив локаторов текста под вопросами от 0 до 7
+    private By[] LocatorsButtonOnMainPage =
+            {
+                    By.xpath(".//div[@class='Header_Nav__AGCXC'] /button[@class='Button_Button__ra12g']"),//кнопка закаазть сверху основной страницы
+                    By.xpath(".//div[@class='Home_RoadMap__2tal_'] /div/button[@class='Button_Button__ra12g Button_Middle__1CSJM']") //нопка заказать снизу страницы
+            }; // массив локаторов кнопок перехода на страницу заказа на основной странице
+
+    private By coockie = By.xpath(".//button[@class='App_CookieButton__3cvqF']");//локатор кнопки куки
     private String[] expectedHeaderTextsQuestionAboutImportant=
             {
                     "Сколько это стоит? И как оплатить?",
@@ -26,7 +47,8 @@ public class MainPage
             "Только начиная с завтрашнего дня. Но скоро станем расторопнее.", "Пока что нет! Но если что-то срочное — всегда можно позвонить в поддержку по красивому номеру 1010.",
             "Самокат приезжает к вам с полной зарядкой. Этого хватает на восемь суток — даже если будете кататься без передышек и во сне. Зарядка не понадобится.",
             "Да, пока самокат не привезли. Штрафа не будет, объяснительной записки тоже не попросим. Все же свои.", "Да, обязательно. Всем самокатов! И Москве, и Московской области."};
-//текст ответов на вопросы о важном
+
+
     public String[] getExpectedHeaderTextsQuestionAboutImportant()
     {
         return expectedHeaderTextsQuestionAboutImportant;
@@ -46,5 +68,21 @@ public class MainPage
     public void clickElement(By element)
     {
         chDriver.findElement(element).click();
+    }
+    public By[] getLocatorsHeaderQuestionAboutImportant()
+    {
+        return locatorsHeaderQuestionAboutImportant;
+    }
+    public By[] getLocatorsBodyQuestionAboutImportant()
+    {
+        return locatorsBodyQuestionAboutImportant;
+    }
+    public By[] getLocatorsButtonMainPage()
+    {
+        return LocatorsButtonOnMainPage;
+    }
+    public By getLocatorCoockie()
+    {
+        return coockie;
     }
 }
