+++ src/test/java/MainPageTest.java
@@ -1,6 +1,4 @@
-import ru.yandex.practicum.Locators.LocatorsMainPage;
-import ru.yandex.practicum.PageObject.MainPage.MainPage;
-import ru.yandex.practicum.PageObject.OrderPage.OrderPage;
+import Pages.MainPage.MainPage;
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.junit.After;
 import org.junit.Assert;
@@ -8,17 +6,12 @@ import org.junit.Before;
 import org.junit.Test;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
-
 import java.util.concurrent.TimeUnit;
-
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
 public class MainPageTest
 {
     private WebDriver chDriver;
-    private LocatorsMainPage locatorsMainPage;
     private MainPage TestMainPage;
-    private OrderPage TestOrderPage;
     private String mainPageUrl = "https://qa-scooter.praktikum-services.ru/";
     private String orderUrl = "https://qa-scooter.praktikum-services.ru/order";
 
@@ -30,12 +23,12 @@ public class MainPageTest
     {
         String[] arrOfActualHeader = new String[8];
         String[] arrOfActualBody = new String[8];
-        for (int i = 0; i< locatorsMainPage.getLocatorsBodyQuestionAboutImportant().length; i++)
+        for (int i = 0; i< TestMainPage.getLocatorsBodyQuestionAboutImportant().length; i++)
         {
-            arrOfActualHeader[i] = TestMainPage.getTextOfElement(locatorsMainPage.getLocatorsHeaderQuestionAboutImportant()[i]);
-            TestMainPage.clickElement(locatorsMainPage.getLocatorsHeaderQuestionAboutImportant()[i]);
-            TestMainPage.waitLoadingElement(locatorsMainPage.getLocatorsBodyQuestionAboutImportant()[i]);
-            arrOfActualBody[i] = TestMainPage.getTextOfElement(locatorsMainPage.getLocatorsBodyQuestionAboutImportant()[i]);
+            arrOfActualHeader[i] = TestMainPage.getTextOfElement(TestMainPage.getLocatorsHeaderQuestionAboutImportant()[i]);
+            TestMainPage.clickElement(TestMainPage.getLocatorsHeaderQuestionAboutImportant()[i]);
+            TestMainPage.waitLoadingElement(TestMainPage.getLocatorsBodyQuestionAboutImportant()[i]);
+            arrOfActualBody[i] = TestMainPage.getTextOfElement(TestMainPage.getLocatorsBodyQuestionAboutImportant()[i]);
 
         }
         Assert.assertArrayEquals(arrOfActualHeader, TestMainPage.getExpectedHeaderTextsQuestionAboutImportant());
@@ -44,12 +37,12 @@ public class MainPageTest
     @Test
     public void checkEnabledButtonsToOrder()
     {
-        chDriver.findElement(locatorsMainPage.getLocatorsButtonMainPage()[0]).click();
+        chDriver.findElement(TestMainPage.getLocatorsButtonMainPage()[0]).click();
         assertEquals(orderUrl, chDriver.getCurrentUrl());
 
         chDriver.navigate().back();
 
-        chDriver.findElement(locatorsMainPage.getLocatorsButtonMainPage()[1]).click();
+        chDriver.findElement(TestMainPage.getLocatorsButtonMainPage()[1]).click();
         assertEquals(orderUrl, chDriver.getCurrentUrl());
     }
     @Before
@@ -58,11 +51,9 @@ public class MainPageTest
         WebDriverManager.chromedriver().setup();
         chDriver = new ChromeDriver();
         chDriver.get(mainPageUrl);
-        locatorsMainPage = new LocatorsMainPage();
         TestMainPage = new MainPage(chDriver);
-        TestOrderPage = new OrderPage(chDriver);
         chDriver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
-        TestMainPage.clickElement(locatorsMainPage.getLocatorCoockie());//кликаем на куки, чтобы оно не мешало
+        TestMainPage.clickElement(TestMainPage.getLocatorCoockie());//кликаем на куки, чтобы оно не мешало
 
     }
     @After
